Beginning of transactional logic

+++ MoveResult.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 namespace ChessersEngine {
     public class MoveAttempt {
         public long playerId { get; set; }
-        public long pieceId { get; set; }
+        public int pieceId { get; set; }
         public long pieceGuid { get; set; }
         public int tileId { get; set; }
     }
